RAG-VERTEX/requirements.txt
google-cloud-aiplatform
streamlit
python-dotenv
pypdf2
google-cloud-storage
pdfplumber
scikit-learn
 